+++ apm-sniffer/optional-plugins/optional-spring-plugins/pom.xml
@@ -32,6 +32,7 @@
         <module>spring-annotation-plugin</module>
         <module>spring-tx-plugin</module>
         <module>optional-spring-cloud</module>
+        <module>spring-webflux-5.x-plugin</module>
     </modules>
 
     <properties>
