+++ test/plugin/scenarios/gateway-2.1.x-scenario/pom.xml
@@ -35,7 +35,7 @@
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <test.framework.version>2.0.2.RELEASE</test.framework.version>
+        <test.framework.version>2.1.2.RELEASE</test.framework.version>
         <docker.image.version>${test.framework.version}</docker.image.version>
     </properties>
