+++ test/plugin/scenarios/webflux-scenario/configuration.yml
@@ -17,4 +17,6 @@
 type: jvm
 entryService: http://localhost:8080/projectA/testcase
 healthCheck: http://localhost:8080/projectA/healthCheck
+runningMode: with_optional
 startScript: ./bin/startup.sh
+withPlugins: apm-spring-webflux-5.x-plugin-*.jar
